+++ src/Public/Utilities.ps1
@@ -38,11 +38,6 @@ function Use-PodeWebTemplates
         $UseHsts
     )
 
-    $mod = (Get-Module -Name Pode -ErrorAction Ignore | Sort-Object -Property Version -Descending | Select-Object -First 1)
-    if (($null -eq $mod) -or ($mod.Version -lt [version]'2.6.0')) {
-        throw "The Pode module is not loaded. You need at least Pode v2.6.0 to use this version of the Pode.Web module."
-    }
-
     if ([string]::IsNullOrWhiteSpace($FavIcon)) {
         $FavIcon = '/pode.web/images/favicon.ico'
     }
@@ -88,6 +83,8 @@ function Use-PodeWebTemplates
             Page = @{
                 Name = 'Home'
                 Path = '/'
+                TItle = 'Home'
+                DisplayName = 'Home'
                 IsSystem = $true
             }
         }